refactor persistence into another file
